[feat]: [CDS-78377] fixing build issues
