+++ 125-cd-nextgen/src/test/resources/recaster/alias/alias.txt
@@ -1281,5 +1281,5 @@ io.harness.cdng.manifest.yaml.summary.ManifestStoreInfo
 io.harness.delegate.beans.instancesync.CustomDeploymentOutcomeMetadata
 io.harness.cdng.pipeline.beans.CustomStageSpecParams
 io.harness.cdng.creator.plan.stage.CustomStageNode
-io.harness.cdng.environment.steps.CustomStageEnvironmentStepParameters
 io.harness.cdng.servicenow.ServiceNowTemplateTypeEnum
+io.harness.cdng.environment.helper.beans.CustomStageEnvironmentStepParameters
